--- exampleJsonFile
@@ -1,7 +1,0 @@
-[{
-"id": 1003,
-"asn": 13367,
-"cachegroup": 29,
-"last_updated": "2012-09-17 15:41:22"
-}
-]
