+++ run
@@ -4,25 +4,18 @@
 source dbInfo
 
 #if "gs" (generate structs) is called or structs folder doesn't exist
-if [ "$1" == "gs" ] || [ "$2" == "gs" ] || [ ! -d "structs" ] ; then
-	echo "" > structs/tmp
-    find structs -type f ! -name 'wrapperStruct.go' -delete
-	go build structGen/structGenerator.go
-	./structGenerator $USERNAME $PASSWORD $ENVIRONMENT
-
-	go build main.go
-	./main $USERNAME $PASSWORD $ENVIRONMENT 1
-elif [ "$1" == "cs" ] || [ "$2" == "cs" ] ; then
-	go build structCustomGenerator.go
-	./structCustomGenerator $USERNAME $PASSWORD $ENVIRONMENT 1
-	go build main.go
-	./main $USERNAME $PASSWORD $ENVIRONMENT 0
+if [ "$1" == "cv" ] || [ "$2" == "cv" ] ; then
+	go build viewGenerator.go
+	./viewGenerator $USERNAME $PASSWORD $ENVIRONMENT 1
+	go build server.go
+	./server $USERNAME $PASSWORD $ENVIRONMENT
 elif [ "$1" == "dv" ] || [ "$2" == "dv" ] ; then
-	go build structCustomGenerator.go
-	./structCustomGenerator $USERNAME $PASSWORD $ENVIRONMENT 0
-	go build main.go
-	./main $USERNAME $PASSWORD $ENVIRONMENT 0
+	go build viewGenerator.go
+	./viewGenerator $USERNAME $PASSWORD $ENVIRONMENT 0
+	
+	go build server.go
+	./server $USERNAME $PASSWORD $ENVIRONMENT
 else
-	go build main.go	
-	./main $USERNAME $PASSWORD $ENVIRONMENT 0
+	go build server.go	
+	./server $USERNAME $PASSWORD $ENVIRONMENT
 fi
